Fix flat_hash_map when used across library boundaries
Singletons don't work well when used across dynamic library boundaries, each library might get its own copy of the singleton.
This means if library A creates a hash table and library B adds elements to it, causing a rehash and therefore deallocation, the `empty_default_table()` won't be recognized correctly because the two libraries have different default tables.

This PR fixes this by treating the default table like any other table, allocating and deallocating it. This causes additional allocations when the hashmap is created or set to empty, so it might not be the optimal solution, but at least it works. A potential better solution might involve nullptr and checking against nullptr.